Better error for set_index with sorted and divisions
When calling `df.set_index(sorted=True, divisions=...)` divisions are
assumed to match the actual splits in the new index column. This means
that there must be the same number of divisions as in the original
frame, and the specified divisions must be valid.

This PR does the following:
- Adds a check that the specified divisions are valid, and throws an
  informative error if they aren't
- Improves the docstring for `set_index` to better reflect this
- Consolidates the `set_index` tests (which were scattered throughout
  the test files), and removes some duplicate tests as found
